hide edit functionality of roles in Account-component
Former-commit-id: dc6e1b05d6e6a3e49bc04420bc39a84b0232794e

+++ app/javascript/bundles/archive/components/UserRole.js
@@ -49,12 +49,16 @@ export default class UserRole extends React.Component {
     }
 
     buttons() {
-        return (
-            <span className={'flyout-sub-tabs-content-ico'}>
-                {this.show()}
-                {this.delete()}
-            </span>
-        )
+        if (!this.props.hideEdit) {
+            return (
+                <span className={'flyout-sub-tabs-content-ico'}>
+                    {this.show()}
+                    {this.delete()}
+                </span>
+            )
+        } else {
+            return null;
+        }
     }
 
     render() {
